Added PCA to DataAnalyzer.
NOT READY FOR DEPLOYMENT.

When box is checked, _every_ time showimg is called, it will add the
current PWOA to the basis, remove the oldest, and perform PCA on the
image. This does not keep track of if it has done PCA already, and it
does not save the basis on closing. Therefore, if you tell it to showimg
on the same image pcaBasisSize times, the basis will simply be a single
PWOA frame copied many times.